apparently, I got the bit wrong and the first fix is wrong, my ccv_cache impl is nearly perfect :-)

+++ lib/ccv.h
@@ -140,7 +140,7 @@ typedef union {
 	struct {
 		uint64_t bitmap;
 		uint64_t set;
-		uint32_t age;
+		uint64_t age;
 	} branch;
 	struct {
 		uint64_t sign;
